requirements.txt
# Core dependencies
requests>=2.25.1
pandas>=1.3.0
numpy>=1.21.0

# Technical analysis
ta>=0.10.0

# Configuration
pyyaml>=5.4.1
python-dotenv>=0.19.0

# AI/LangChain (with version constraints for compatibility)
langchain>=0.1.0,<0.2.0
langchain-anthropic>=0.1.0,<0.2.0
pydantic>=1.10.0,<2.0.0

# Output formatting
tabulate>=0.9.0

# Optional: Telegram notifications
python-telegram-bot>=20.0

# Optional: Additional utilities
python-dateutil>=2.8.0
